requirements.txt
streamlit
pytrends
praw
transformers
torch
deep-translator
moviepy
gTTS
edge-tts
supabase
pandas
requests
google-genai
python-dotenv
regex
Pillow
pypdf
python-docx
altair
